+++ gtk2/common/common.h
@@ -659,9 +659,6 @@ typedef struct {
                      highlightTab,
                      roundAllTabs,
                      animatedProgress,
-#ifdef QTC_GTK2_ENABLE_PARENTLESS_DIALOG_FIX_SUPPORT
-                     fixParentlessDialogs,
-#endif
                      customMenuTextColor,
                      menubarMouseOver,
                      useHighlightForMenu,
@@ -794,9 +791,6 @@ typedef struct {
                      noBgndOpacityApps,
                      noMenuBgndOpacityApps,
                      noBgndImageApps;
-#ifdef QTC_GTK2_ENABLE_PARENTLESS_DIALOG_FIX_SUPPORT
-    Strings          noDlgFixApps;
-#endif
     Strings          noMenuStripeApps;
 } Options;